requirements.txt
flask==2.3.3
PyPDF2==3.0.1
pandas==2.1.1
xlsxwriter==3.1.9
openai==0.28.1
python-dotenv==1.0.0
argparse==1.4.0 